requirements.txt
Flask==2.2.5
requests==2.31.0
Werkzeug==2.2.3
pysmb==1.2.9.1
